Refactoring to split connection and api
