+++ chain-connect/src/chainApis/index.ts
@@ -12,7 +12,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from "./customClients";
-export * from "./GalachainClient";
-export * from "./utils/WalletUtils";
-export * from "./chainApis";
+export * from "./TokenApi";
+export * from "./PublicKeyApi";
